refactor: move addBinding to tests dir
used only there
